// src/auth/interfaces/valid-permissions.interface.ts
export enum ValidPermissions {
  VER_USUARIOS = 'VER_USUARIOS',
  CREAR_USUARIO = 'CREAR_USUARIO',
  EDITAR_USUARIO = 'EDITAR_USUARIO',
  ELIMINAR_USUARIO = 'ELIMINAR_USUARIO',
  VER_PRODUCTOS = 'VER_PRODUCTOS',
  CREAR_PRODUCTO = 'CREAR_PRODUCTO',
  EDITAR_PRODUCTO = 'EDITAR_PRODUCTO',
  ELIMINAR_PRODUCTO = 'ELIMINAR_PRODUCTO',
  VER_CLIENTES = 'VER_CLIENTES',
  CREAR_CLIENTE = 'CREAR_CLIENTE',
  EDITAR_CLIENTE = 'EDITAR_CLIENTE',
  ELIMINAR_CLIENTE = 'ELIMINAR_CLIENTE',
  VER_MARCAS = 'VER_MARCAS',
  CREAR_MARCA = 'CREAR_MARCA',
  EDITAR_MARCA = 'EDITAR_MARCA',
  ELIMINAR_MARCA = 'ELIMINAR_MARCA',
  VER_LINEAS = 'VER_LINEAS',
  CREAR_LINEA = 'CREAR_LINEA',
  EDITAR_LINEA = 'EDITAR_LINEA',
  ELIMINAR_LINEA = 'ELIMINAR_LINEA',
  VER_GRUPOS = 'VER_GRUPOS',
  CREAR_GRUPO = 'CREAR_GRUPO',
  EDITAR_GRUPO = 'EDITAR_GRUPO',
  ELIMINAR_GRUPO = 'ELIMINAR_GRUPO',
  VER_PERFILES = 'VER_PERFILES',
  CREAR_PERFIL = 'CREAR_PERFIL',
  EDITAR_PERFIL = 'EDITAR_PERFIL',
  ELIMINAR_PERFIL = 'ELIMINAR_PERFIL',
  VER_VENDEDORES = 'VER_VENDEDORES',
  CREAR_VENDEDOR = 'CREAR_VENDEDOR',
  EDITAR_VENDEDOR = 'EDITAR_VENDEDOR',
  ELIMINAR_VENDEDOR = 'ELIMINAR_VENDEDOR',
  VER_PROVEEDORES = 'VER_PROVEEDORES',
  CREAR_PROVEEDOR = 'CREAR_PROVEEDOR',
  EDITAR_PROVEEDOR = 'EDITAR_PROVEEDOR',
  ELIMINAR_PROVEEDOR = 'ELIMINAR_PROVEEDOR',
  VER_COMPRAS = 'VER_COMPRAS',
  CREAR_COMPRA = 'CREAR_COMPRA',
  EDITAR_COMPRA = 'EDITAR_COMPRA',
  ELIMINAR_COMPRA = 'ELIMINAR_COMPRA',
  VER_PEDIDOS = 'VER_PEDIDOS',
  CREAR_PEDIDO = 'CREAR_PEDIDO',
  EDITAR_PEDIDO = 'EDITAR_PEDIDO',
  ELIMINAR_PEDIDO = 'ELIMINAR_PEDIDO',
}